backend/requirements.txt
# Web Framework
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9

# Database
sqlalchemy==2.0.27
asyncpg==0.29.0
alembic==1.13.1
psycopg2-binary==2.9.9

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
firebase-admin==6.5.0
pydantic[email]==2.6.1
pydantic-settings==2.1.0

# Utilities
python-dotenv==1.0.1
httpx==0.27.0
Pillow==10.4.0
aiofiles==23.2.1

# Validation
email-validator==2.1.0

# Development
pytest==8.0.2
pytest-asyncio==0.23.5
black==24.1.1
flake8==7.0.0